+++ src/providers/Flatworld.php
@@ -225,8 +225,7 @@ class Flatworld extends Provider
             $msg = "{$method} :: {$uniqueId} :: {$message}";
         }
 
-        $logger = new Logger(true);
-        return $logger->logMessage($msg);
+        return Logger::logMessage($msg, true);
     }
 
     /**
@@ -241,6 +240,7 @@ class Flatworld extends Provider
     {
         $file = 'NA';
         $line = 'NA';
+        $devMode = Craft::$app->getConfig()->getGeneral()->devMode;
 
         if (method_exists($error, 'hasResponse')) {
             $data = Json::decode((string) $error->getResponse()->getBody());
@@ -284,7 +284,7 @@ class Flatworld extends Provider
 
         $this->_logMessage(__METHOD__, $debugMessage, $uniqueId);
 
-        if ($this->getSetting('enableErrorEmailMessages')) {
+        if (!$devMode && $this->getSetting('enableErrorEmailMessages')) {
             $mailer = new Mailer($this->getSetting('displayDebugMessages'));
             $mailer->sendMail($debugMessage);
         }