+++ src/services/Mailer.php
@@ -16,6 +16,9 @@ use fireclaytile\flatworld\services\Logger;
  */
 class Mailer extends Component
 {
+    const MAIL_TO = 'web@fireclaytile.com';
+    const MAIL_SUBJECT = 'Flatworld Plugin - Error Found';
+
     /**
      * @var boolean|null
      */
@@ -41,16 +44,18 @@ class Mailer extends Component
     {
         try {
             $message = new Message();
-            $message->setTo('web@fireclaytile.com');
-            $message->setSubject('Flatworld Plugin - Error Found');
+            $message->setTo(self::MAIL_TO);
+            $message->setSubject(self::MAIL_SUBJECT);
             $message->setTextBody($textBody);
 
             Craft::$app->getMailer()->send($message);
         } catch (Throwable $e) {
-            $logger = new Logger($this->_loggingEnabled);
-            $logger->logMessage(
-                __METHOD__ . ' :: Something went wrong: ' . $e->getMessage(),
-            );
+            $msg =
+                __METHOD__ .
+                ' :: Something went wrong: ' .
+                $e->getMessage() .
+                '';
+            Logger::logMessage($msg, $this->_loggingEnabled);
         }
     }
 }