+++ tests/unit/RatesServiceTest.php
@@ -1248,8 +1248,7 @@ class RatesServiceTest extends Unit
 
         $msg = "{$method} :: {$message}";
 
-        $logger = new Logger(true);
-        return $logger->logMessage($msg);
+        return Logger::logMessage($msg, true);
     }
 
     /**